ig 일반 페이지 코드 통일 (#737)
* ig 일반 페이지 코드 통일

* 리뷰 수정
