+++ src/app/(ig)/components/detail/EditIgButton.jsx
