+++ src/app/(ig)/components/detail/IgContents.jsx
