+++ src/app/(ig)/components/detail/IgDetail.module.css
