+++ src/app/(ig)/pig/[id]/loading.jsx
