+++ src/app/(ig)/pig/create/page.jsx
@@ -1,0 +1,7 @@
+import CreateIgPage from '@/app/(ig)/components/editor/CreateIgPage';
+
+export const metadata = { title: 'PIG' };
+
+export default async function CreatePigPage() {
+  return <CreateIgPage kind="pig" />;
+}
