+++ src/app/(ig)/sig/create/page.jsx
@@ -1,0 +1,7 @@
+import CreateIgPage from '@/app/(ig)/components/editor/CreateIgPage';
+
+export const metadata = { title: 'SIG' };
+
+export default async function CreateSigPage() {
+  return <CreateIgPage kind="sig" />;
+}
